+++ convex/line/mutations.ts
@@ -1,6 +1,5 @@
 import { ConvexError, v } from "convex/values";
 import { internal } from "../_generated/api";
-import type { Id } from "../_generated/dataModel";
 import { internalMutation } from "../_generated/server";
 import { managerMutation } from "../_lib/functions";
 import { buildLineAuthorizeUrl } from "../_lib/lineClient";
@@ -275,39 +274,5 @@ export const sendInvite = managerMutation({
     await ctx.scheduler.runAfter(0, internal.line.actions.sendInviteEmail, {
       staffId: staff._id,
     });
-  },
-});
-
-/**
- * 一括: 未連携 or 友達解除のスタッフ全員に依頼メール
- *
- * - rate limit は `lineInviteBulk`（3回/時、shopId キー）。個別ボタンとは別バケット
- * - Resend / LINE 側のスロットリングを避けるため、scheduler で 100ms 間隔に分散
- */
-const BULK_INVITE_INTERVAL_MS = 100;
-
-export const sendInviteBulk = managerMutation({
-  args: {},
-  handler: async (ctx) => {
-    const { ok } = await rateLimit(ctx, {
-      name: "lineInviteBulk",
-      key: ctx.shop._id,
-    });
-    if (!ok) {
-      throw new ConvexError("送信が集中しています。しばらく待ってからお試しください");
-    }
-
-    const staffs = await ctx.db
-      .query("staffs")
-      .withIndex("by_shopId_isDeleted", (q) => q.eq("shopId", ctx.shop._id).eq("isDeleted", false))
-      .collect();
-    const targets = staffs.filter((s) => s.email && (!s.lineUserId || s.lineFollowing === false));
-
-    for (let i = 0; i < targets.length; i++) {
-      await ctx.scheduler.runAfter(i * BULK_INVITE_INTERVAL_MS, internal.line.actions.sendInviteEmail, {
-        staffId: targets[i]._id as Id<"staffs">,
-      });
-    }
-    return { sentCount: targets.length };
   },
 });
