+++ convex/line/queries.ts
@@ -46,22 +46,6 @@ export const getQuotaStatus = managerQuery({
 });
 
 /**
- * 店舗内で一括連携依頼の対象になるスタッフ数（店長UI用）
- */
-export const getBulkInviteTargetCount = managerQuery({
-  args: {},
-  handler: async (ctx) => {
-    const shop = ctx.shop;
-    if (!shop) return 0;
-    const staffs = await ctx.db
-      .query("staffs")
-      .withIndex("by_shopId_isDeleted", (q) => q.eq("shopId", shop._id).eq("isDeleted", false))
-      .take(1000);
-    return staffs.filter((s) => s.email && (!s.lineUserId || s.lineFollowing === false)).length;
-  },
-});
-
-/**
  * lineUserId からスタッフを引く（Webhook で使う）
  */
 export const findStaffByLineUserId = internalQuery({
@@ -104,29 +88,6 @@ export const getInviteEmailData = internalQuery({
       staffName: staff.name,
       staffEmail: staff.email,
       shopName: shop.name,
-    };
-  },
-});
-
-/**
- * 一括連携依頼メールの対象スタッフ取得
- * 未連携 or 友達解除されているスタッフのみ
- */
-export const getBulkInviteTargets = internalQuery({
-  args: { shopId: v.id("shops") },
-  handler: async (ctx, { shopId }) => {
-    const shop = await ctx.db.get(shopId);
-    if (!shop || shop.isDeleted) return null;
-    const staffs = await ctx.db
-      .query("staffs")
-      .withIndex("by_shopId_isDeleted", (q) => q.eq("shopId", shopId).eq("isDeleted", false))
-      .collect();
-    return {
-      shopName: shop.name,
-      shopId,
-      targets: staffs
-        .filter((s) => s.email && (!s.lineUserId || s.lineFollowing === false))
-        .map((s) => ({ staffId: s._id, name: s.name, email: s.email })),
     };
   },
 });
