+++ doc/features/line-notification.md
@@ -25,19 +25,17 @@
 - `src/routes/_unregistered/line.callback.tsx` — OAuth コールバックページ
 - `src/components/features/Line/LineLinkQrDialog/` — 店長UI: QR / URL 表示
 - `src/components/features/Line/LineInviteConfirmContent/` — 個別連携依頼確認モーダル中身
-- `src/components/features/Line/LineBulkInviteContent/` — 一括連携依頼確認モーダル中身
 - `src/components/features/Line/LineCallbackPage/` — コールバック完了 / エラー UI
 - `src/components/features/Dashboard/StaffRoster/StaffRow.tsx` — `…` メニューに LINE 連携項目追加
-- `src/components/features/Dashboard/StaffRoster/index.tsx` — 上部に「未連携全員に依頼」ボタン
 - `src/components/features/Dashboard/DashboardContent/index.tsx` — モーダル接続
 
 ## 画面一覧
 
 | 画面 | 役割 |
 |---|---|
-| 店長ダッシュボード（既存）| StaffRow メニュー / 一括ボタン経由で連携依頼 |
+| 店長ダッシュボード（既存）| StaffRow メニュー経由で連携リンク表示 / 個別連携依頼 |
 | LineLinkQrDialog | QR 表示 + URL コピー |
-| 連携依頼確認ダイアログ（個別 / 一括） | 送信前の確認 |
+| 連携依頼確認ダイアログ（個別） | 送信前の確認 |
 | `/line/callback` | OAuth 完了画面（成功 / 期限切れ / レート超過 / エラー） |
 | LINE 公式アカウントトーク画面 | 受信メッセージへ Reply API で定型応答 |
 
@@ -47,7 +45,6 @@
 |---|---|---|
 | `api.line.mutations.generateLinkToken` | mutation | 店長UIから連携用QR/URL発行 |
 | `api.line.mutations.sendInvite` | mutation | 個別スタッフへ連携依頼メール |
-| `api.line.mutations.sendInviteBulk` | mutation | 未連携全員へ連携依頼メール |
 | `api.line.queries.getLinkStatusByShop` | query | 店舗のスタッフごと連携状況 |
 | `api.line.queries.getQuotaStatus` | query | Quota 状態（normal / exceeded） |
 | `api.line.actions.redeemLineToken` | action | OAuth コールバック処理（state 検証 → code 交換 → 連携完了） |
@@ -73,7 +70,6 @@
 - `lineLinkRedeem`: 5回/分（state先頭8文字キー） — OAuth コールバックのブルートフォース防御
 - `lineWebhook`: 100回/分（global） — Webhook 暴発時のセーフティネット
 - `lineInvite`: 30回/時（shopId キー） — 個別連携依頼メールの連打防止
-- `lineInviteBulk`: 3回/時（shopId キー） — 一括連携依頼メールの連打防止（一括送信は 100ms 間隔で scheduler 分散）
 
 ## 環境変数
 
@@ -88,9 +84,10 @@
 
 ## 追加スタッフへの募集中通知
 
-シフト募集中にスタッフを追加した場合、追加スタッフにも希望提出リンクをメールで送る。LINEログイン完了時に友だち追加済み、またはWebhook followで `lineFollowing` が `true` になった場合は、同じ対象募集の希望提出リンクをLINEで送る。
+スタッフ追加時に、スタッフ向け利用規約/プライバシーポリシー同意依頼メールとは別に LINE 連携依頼メールも送る。シフト募集中にスタッフを追加した場合、追加スタッフにも希望提出リンクをメールで送る。LINEログイン完了時に友だち追加済み、またはWebhook followで `lineFollowing` が `true` になった場合は、同じ対象募集の希望提出リンクをLINEで送る。
 
 - 対象募集: `status === "open"`、未削除、締切前または締切当日
+- LINE連携依頼メール: `staff.mutations.addStaffs` から追加スタッフごとに `internal.line.actions.sendInviteEmail` をスケジュール
 - メール通知: `staff.mutations.addStaffs` から追加スタッフごとに `internal.email.actions.sendOpenRecruitmentNotificationEmailsForStaff` をスケジュール
 - LINE通知: `line.mutations.finalizeLinking` / `dispatchWebhookEvents` から `internal.email.actions.sendOpenRecruitmentNotificationLinesForStaff` をスケジュール
 - 複数の対象募集がある場合は募集ごとに1通ずつ送る
